feat(session): 统计指标改为活动计数 + 派生表，并修掉 PG 时间戳溢出与页面滚动
- 统计指标：把「事件」换成轮次 / 步骤 / 用户输入 / 工具调用（按 t_events.f_type 数）；
  totals / subagent / human 改由 host 按是否子代理聚合一次；按模型的行只带 token 用量
  （活动计数没有模型归属）
- 活动计数落派生表 t_event_counts（故意不进 drizzle schema / 迁移，建表幂等、可销毁重建）：
  写路径在事务提交后旁路累加（失败只 warn）、表为空时启动回填、rewind / fork 按会话重算、
  删除与 GC 清行；读侧只读这张表，不再逐事件数类型
- 修掉 t_event_usage.f_created_at 在 PG 侧是 integer（int4）导致毫秒时间戳溢出 22003：
  列类型改为 bigint（TableDef + 迁移 DDL；未新增迁移、未 bump SCHEMA_VERSION，
  已存在的 PG 库需手工 ALTER 或重建）
- 对话管理页：页头（标题 / tab / 动作）与搜索行固定在顶部，只有内容区滚动；
  列表刷新入口改为可选探测（refresh 在运行时不一定存在）
- 记录同步：表结构（派生统计表）/ 写路径 / 用量统计 ADR / how-to-verify

验证：pnpm exec vitest run packages/session 501 passed；just pg test 36 passed；
oxlint packages/session 0 error；session-rdb 与 ui-conversation-manager build 通过

+++ packages/session/session-rdb/drizzle/postgres/20260918120000_v3_event_usage/migration.sql
@@ -1,6 +1,6 @@
 CREATE TABLE "t_event_usage" (
 	"f_event_id" text PRIMARY KEY NOT NULL,
-	"f_created_at" integer NOT NULL,
+	"f_created_at" bigint NOT NULL,
 	"f_provider" text,
 	"f_model" text,
 	"f_input_tokens" integer DEFAULT 0 NOT NULL,
